pyproject.toml
[tool.poetry]
name = "finance"
version = "0.1.0"
description = ""
authors = ["noam.cohen <noam.cohen@explorium.ai>"]

[tool.poetry.dependencies]
python = "3.9"
selenium = "^4.1.3"
pandas = "^1.4.2"
prefect = "^2.0.2"
pytest = "^7.1.2"
pyppeteer = "^1.0.2"
python-dotenv = "^0.20.0"
pymongo = "^4.1.1"
SQLAlchemy = "^1.4.37"
PyMySQL = "^1.0.2"
gcsfs = "^2022.7.1"
google-search-results = "^2.4.1"
wordcloud = "^1.8.2"
pandas-profiling = "^3.3.0"
catboost = "^1.0.6"
scikit-learn = "^1.1.2"
deepchecks = "^0.9.0"
shap = "^0.41.0"
clearml = "^1.8.1"
Deprecated = "^1.2.13"
omegaconf = "^2.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
